#775 Command request, context, and parsing exceptions
