+++ hartshorn-commands/src/main/java/org/dockbox/hartshorn/commands/CommandListenerImpl.java
@@ -16,7 +16,6 @@
 
 package org.dockbox.hartshorn.commands;
 
-import org.dockbox.hartshorn.commands.exceptions.ParsingException;
 import org.dockbox.hartshorn.application.context.ApplicationContext;
 
 import java.io.IOException;
