+++ hartshorn-commands/src/main/java/org/dockbox/hartshorn/commands/context/IllegalArgumentRequestException.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2019-2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.dockbox.hartshorn.commands.context;
+
+import org.dockbox.hartshorn.util.ApplicationRuntimeException;
+
+public class IllegalArgumentRequestException extends ApplicationRuntimeException {
+    public IllegalArgumentRequestException(final String message) {
+        super(message);
+    }
+}
